Adding auth to tweet management

+++ app/api/tweets.py
@@ -4,6 +4,7 @@ import json
 from flask import Blueprint, jsonify, request
 from app import db
 from app.models import Tweet
+from app.utils import validate_token
 from app.schemas import tweets_schema, tweet_schema
 from datetime import datetime
 
@@ -23,13 +24,18 @@ def get_tweet(id):
 
 @api.route('/tweets', methods=['POST'])
 def post_tweet():
+    if 'Authorization' not in request.headers:
+        return jsonify({"error" : "Not authorized"}), 403
+    user = validate_token(request.headers['Authorization'])
+    if user is None:
+        return jsonify({"error" : "Not authorized"}), 403
     try:
         payload = json.loads(request.data)
     except ValueError :
         return jsonify({"error" : "Bad payload received"}), 422
     if 'text' not in payload:
         return jsonify({"error" : "Bad payload received"}), 422
-    tweet = Tweet(text=payload['text'], created_at=datetime.now())
+    tweet = Tweet(text=payload['text'], created_at=datetime.now(), user_id=user.id)
     db.session.add(tweet)
     db.session.flush()
     db.session.commit()
@@ -37,7 +43,14 @@ def post_tweet():
 
 @api.route('/tweets/<int:id>', methods=['DELETE'])
 def delete_tweet(id):
+    if 'Authorization' not in request.headers:
+        return jsonify({"error" : "Not authorized"}), 403
+    user = validate_token(request.headers['Authorization'])
+    if user is None:
+        return jsonify({"error" : "Not authorized"}), 403
     tweet = db.session.query(Tweet).get(id)
+    if user.id != tweet.user_id:
+        return jsonify({"error" : "Not authorized (not your tweet)"}), 403
     if tweet is not None :
         db.session.delete(tweet)
         db.session.flush()
@@ -47,6 +60,11 @@ def delete_tweet(id):
 
 @api.route('/tweets/<int:id>', methods=['PATCH'])
 def change_tweet(id):
+    if 'Authorization' not in request.headers:
+        return jsonify({"error" : "Not authorized"}), 403
+    user = validate_token(request.headers['Authorization'])
+    if user is None:
+        return jsonify({"error" : "Not authorized"}), 403
     try:
         payload = json.loads(request.data)
     except ValueError :
@@ -54,6 +72,8 @@ def change_tweet(id):
     if 'text' not in payload:
         return jsonify({"error" : "Bad payload received"}), 422
     tweet = db.session.query(Tweet).get(id)
+    if user.id != tweet.user_id:
+        return jsonify({"error" : "Not authorized (not your tweet)"}), 403
     if tweet is not None :
         tweet.text = payload['text']
         db.session.commit()
